refactor: UI changes suggested by users.
