+++ src/components/ArenaHeader.tsx
@@ -17,7 +17,7 @@ const ArenaHeader: React.FC<ArenaHeaderProps> = ({ challengeName, subtitle, judg
             <div><svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" height="24" width="24" className="pointer-events-none size-3.5 select-none sm:size-4 scale-100"><path d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z" fill="#4285F4"></path><path d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z" fill="#34A853"></path><path d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z" fill="#FBBC05"></path><path d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z" fill="#EA4335"></path><path d="M1 1h22v22H1z" fill="none"></path></svg></div>
             <div><svg xmlns="http://www.w3.org/2000/svg" stroke-linejoin="round" viewBox="0 0 16 16" height="24" width="24" className="fill-orange-400 pointer-events-none size-3.5 select-none sm:size-4 scale-100"><path d="M14.9449 6.54871C15.3128 5.45919 15.1861 4.26567 14.5978 3.27464C13.7131 1.75461 11.9345 0.972595 10.1974 1.3406C9.42464 0.481584 8.3144 -0.00692594 7.15045 7.42132e-05C5.37487 -0.00392587 3.79946 1.1241 3.2532 2.79113C2.11256 3.02164 1.12799 3.72615 0.551837 4.72468C-0.339497 6.24071 -0.1363 8.15175 1.05451 9.45178C0.686626 10.5413 0.813308 11.7348 1.40162 12.7258C2.28637 14.2459 4.06498 15.0279 5.80204 14.6599C6.5743 15.5189 7.68504 16.0074 8.849 15.9999C10.6256 16.0044 12.2015 14.8754 12.7478 13.2069C13.8884 12.9764 14.873 12.2718 15.4491 11.2733C16.3394 9.75728 16.1357 7.84774 14.9454 6.54771L14.9449 6.54871ZM8.85001 14.9544C8.13907 14.9554 7.45043 14.7099 6.90468 14.2604C6.92951 14.2474 6.97259 14.2239 7.00046 14.2069L10.2293 12.3668C10.3945 12.2743 10.4959 12.1008 10.4949 11.9133V7.42173L11.8595 8.19925C11.8742 8.20625 11.8838 8.22025 11.8858 8.23625V11.9558C11.8838 13.6099 10.5263 14.9509 8.85001 14.9544ZM2.32133 12.2028C1.9651 11.5958 1.8369 10.8843 1.95902 10.1938C1.98284 10.2078 2.02489 10.2333 2.05479 10.2503L5.28366 12.0903C5.44733 12.1848 5.65003 12.1848 5.81421 12.0903L9.75604 9.84429V11.3993C9.75705 11.4153 9.74945 11.4308 9.73678 11.4408L6.47295 13.3004C5.01915 14.1264 3.1625 13.6354 2.32184 12.2028H2.32133ZM1.47155 5.24819C1.82626 4.64017 2.38619 4.17516 3.05305 3.93366C3.05305 3.96116 3.05152 4.00966 3.05152 4.04366V7.72424C3.05051 7.91124 3.15186 8.08475 3.31654 8.17725L7.25838 10.4228L5.89376 11.2003C5.88008 11.2093 5.86285 11.2108 5.84765 11.2043L2.58331 9.34327C1.13255 8.51426 0.63494 6.68272 1.47104 5.24869L1.47155 5.24819ZM12.6834 7.82274L8.74157 5.57669L10.1062 4.79968C10.1199 4.79068 10.1371 4.78918 10.1523 4.79568L13.4166 6.65522C14.8699 7.48373 15.3681 9.31827 14.5284 10.7523C14.1732 11.3593 13.6138 11.8243 12.9474 12.0663V8.27575C12.9489 8.08875 12.8481 7.91574 12.6839 7.82274H12.6834ZM14.0414 5.8057C14.0176 5.7912 13.9756 5.7662 13.9457 5.7492L10.7168 3.90916C10.5531 3.81466 10.3504 3.81466 10.1863 3.90916L6.24442 6.15521V4.60017C6.2434 4.58417 6.251 4.56867 6.26367 4.55867L9.52751 2.70063C10.9813 1.87311 12.84 2.36563 13.6781 3.80066C14.0323 4.40667 14.1605 5.11618 14.0404 5.8057H14.0414ZM5.50257 8.57726L4.13744 7.79974C4.12275 7.79274 4.11312 7.77874 4.11109 7.76274V4.04316C4.11211 2.38713 5.47368 1.0451 7.15197 1.0461C7.86189 1.0461 8.54902 1.2921 9.09476 1.74011C9.06993 1.75311 9.02737 1.77661 8.99899 1.79361L5.77012 3.63365C5.60493 3.72615 5.50358 3.89916 5.50459 4.08666L5.50257 8.57626V8.57726ZM6.24391 7.00022L7.99972 5.9997L9.75553 6.99972V9.00027L7.99972 10.0003L6.24391 9.00027V7.00022Z"></path></svg></div>
             <div><svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" className="pointer-events-none size-3.5 select-none sm:size-4 scale-125" data-sentry-element="svg" data-sentry-component="Meta" data-sentry-source-file="meta.tsx"><title>Meta</title><path d="M6.897 4h-.024l-.031 2.615h.022c1.715 0 3.046 1.357 5.94 6.246l.175.297.012.02 1.62-2.438-.012-.019a48.763 48.763 0 00-1.098-1.716 28.01 28.01 0 00-1.175-1.629C10.413 4.932 8.812 4 6.896 4z" fill="url(#lobe-icons-meta-fill-0)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M6.873 4C4.95 4.01 3.247 5.258 2.02 7.17a4.352 4.352 0 00-.01.017l2.254 1.231.011-.017c.718-1.083 1.61-1.774 2.568-1.785h.021L6.896 4h-.023z" fill="url(#lobe-icons-meta-fill-1)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M2.019 7.17l-.011.017C1.2 8.447.598 9.995.274 11.664l-.005.022 2.534.6.004-.022c.27-1.467.786-2.828 1.456-3.845l.011-.017L2.02 7.17z" fill="url(#lobe-icons-meta-fill-2)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M2.807 12.264l-2.533-.6-.005.022c-.177.918-.267 1.851-.269 2.786v.023l2.598.233v-.023a12.591 12.591 0 01.21-2.44z" fill="url(#lobe-icons-meta-fill-3)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M2.677 15.537a5.462 5.462 0 01-.079-.813v-.022L0 14.468v.024a8.89 8.89 0 00.146 1.652l2.535-.585a4.106 4.106 0 01-.004-.022z" fill="url(#lobe-icons-meta-fill-4)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M3.27 16.89c-.284-.31-.484-.756-.589-1.328l-.004-.021-2.535.585.004.021c.192 1.01.568 1.85 1.106 2.487l.014.017 2.018-1.745a2.106 2.106 0 01-.015-.016z" fill="url(#lobe-icons-meta-fill-5)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M10.78 9.654c-1.528 2.35-2.454 3.825-2.454 3.825-2.035 3.2-2.739 3.917-3.871 3.917a1.545 1.545 0 01-1.186-.508l-2.017 1.744.014.017C2.01 19.518 3.058 20 4.356 20c1.963 0 3.374-.928 5.884-5.33l1.766-3.13a41.283 41.283 0 00-1.227-1.886z" fill="#0082FB" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M13.502 5.946l-.016.016c-.4.43-.786.908-1.16 1.416.378.483.768 1.024 1.175 1.63.48-.743.928-1.345 1.367-1.807l.016-.016-1.382-1.24z" fill="url(#lobe-icons-meta-fill-6)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M20.918 5.713C19.853 4.633 18.583 4 17.225 4c-1.432 0-2.637.787-3.723 1.944l-.016.016 1.382 1.24.016-.017c.715-.747 1.408-1.12 2.176-1.12.826 0 1.6.39 2.27 1.075l.015.016 1.589-1.425-.016-.016z" fill="#0082FB" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M23.998 14.125c-.06-3.467-1.27-6.566-3.064-8.396l-.016-.016-1.588 1.424.015.016c1.35 1.392 2.277 3.98 2.361 6.971v.023h2.292v-.022z" fill="url(#lobe-icons-meta-fill-7)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M23.998 14.15v-.023h-2.292v.022c.004.14.006.282.006.424 0 .815-.121 1.474-.368 1.95l-.011.022 1.708 1.782.013-.02c.62-.96.946-2.293.946-3.91 0-.083 0-.165-.002-.247z" fill="url(#lobe-icons-meta-fill-8)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M21.344 16.52l-.011.02c-.214.402-.519.67-.917.787l.778 2.462a3.493 3.493 0 00.438-.182 3.558 3.558 0 001.366-1.218l.044-.065.012-.02-1.71-1.784z" fill="url(#lobe-icons-meta-fill-9)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M19.92 17.393c-.262 0-.492-.039-.718-.14l-.798 2.522c.449.153.927.222 1.46.222.492 0 .943-.073 1.352-.215l-.78-2.462c-.167.05-.341.075-.517.073z" fill="url(#lobe-icons-meta-fill-10)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M18.323 16.534l-.014-.017-1.836 1.914.016.017c.637.682 1.246 1.105 1.937 1.337l.797-2.52c-.291-.125-.573-.353-.9-.731z" fill="url(#lobe-icons-meta-fill-11)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><path d="M18.309 16.515c-.55-.642-1.232-1.712-2.303-3.44l-1.396-2.336-.011-.02-1.62 2.438.012.02.989 1.668c.959 1.61 1.74 2.774 2.493 3.585l.016.016 1.834-1.914a2.353 2.353 0 01-.014-.017z" fill="url(#lobe-icons-meta-fill-12)" data-sentry-element="path" data-sentry-source-file="meta.tsx"></path><defs data-sentry-element="defs" data-sentry-source-file="meta.tsx"><linearGradient id="lobe-icons-meta-fill-0" x1="75.897%" x2="26.312%" y1="89.199%" y2="12.194%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset=".06%" stop-color="#0867DF" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="45.39%" stop-color="#0668E1" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="85.91%" stop-color="#0064E0" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-1" x1="21.67%" x2="97.068%" y1="75.874%" y2="23.985%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="13.23%" stop-color="#0064DF" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="99.88%" stop-color="#0064E0" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-2" x1="38.263%" x2="60.895%" y1="89.127%" y2="16.131%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="1.47%" stop-color="#0072EC" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="68.81%" stop-color="#0064DF" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-3" x1="47.032%" x2="52.15%" y1="90.19%" y2="15.745%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="7.31%" stop-color="#007CF6" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="99.43%" stop-color="#0072EC" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-4" x1="52.155%" x2="47.591%" y1="58.301%" y2="37.004%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="7.31%" stop-color="#007FF9" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="100%" stop-color="#007CF6" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-5" x1="37.689%" x2="61.961%" y1="12.502%" y2="63.624%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="7.31%" stop-color="#007FF9" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="100%" stop-color="#0082FB" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-6" x1="34.808%" x2="62.313%" y1="68.859%" y2="23.174%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="27.99%" stop-color="#007FF8" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="91.41%" stop-color="#0082FB" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-7" x1="43.762%" x2="57.602%" y1="6.235%" y2="98.514%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="0%" stop-color="#0082FB" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="99.95%" stop-color="#0081FA" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-8" x1="60.055%" x2="39.88%" y1="4.661%" y2="69.077%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="6.19%" stop-color="#0081FA" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="100%" stop-color="#0080F9" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-9" x1="30.282%" x2="61.081%" y1="59.32%" y2="33.244%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="0%" stop-color="#027AF3" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="100%" stop-color="#0080F9" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-10" x1="20.433%" x2="82.112%" y1="50.001%" y2="50.001%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="0%" stop-color="#0377EF" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="99.94%" stop-color="#0279F1" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-11" x1="40.303%" x2="72.394%" y1="35.298%" y2="57.811%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset=".19%" stop-color="#0471E9" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="100%" stop-color="#0377EF" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient><linearGradient id="lobe-icons-meta-fill-12" x1="32.254%" x2="68.003%" y1="19.719%" y2="84.908%" data-sentry-element="linearGradient" data-sentry-source-file="meta.tsx"><stop offset="27.65%" stop-color="#0867DF" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop><stop offset="100%" stop-color="#0471E9" data-sentry-element="stop" data-sentry-source-file="meta.tsx"></stop></linearGradient></defs></svg></div>
-            <div><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 18 18" fill="none" className="fill-white pointer-events-none size-3.5 select-none sm:size-4 scale-90"><path d="M14.0898 5.7984L14.4005 18.0004H16.889L17.2001 1.35645L14.0898 5.7984Z"></path><path d="M17.2007 0H13.4037L7.44531 8.51004L9.34383 11.2214L17.2007 0Z"></path><path d="M0.799805 18.0001H4.59683L6.49572 15.2888L4.59683 12.5771L0.799805 18.0001Z"></path><path d="M0.799805 5.79785L9.34349 17.9999H13.1405L4.59683 5.79785H0.799805Z"></path></svg></div>
+            <div><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 18 18" fill="none" className="fill-neutral-900 dark:fill-neutral-100 pointer-events-none size-3.5 select-none sm:size-4 scale-90"><path d="M14.0898 5.7984L14.4005 18.0004H16.889L17.2001 1.35645L14.0898 5.7984Z"></path><path d="M17.2007 0H13.4037L7.44531 8.51004L9.34383 11.2214L17.2007 0Z"></path><path d="M0.799805 18.0001H4.59683L6.49572 15.2888L4.59683 12.5771L0.799805 18.0001Z"></path><path d="M0.799805 5.79785L9.34349 17.9999H13.1405L4.59683 5.79785H0.799805Z"></path></svg></div>
             <div><svg width="24" height="24" viewBox="0 0 191 135" fill="none" xmlns="http://www.w3.org/2000/svg" className="pointer-events-none size-3.5 select-none sm:size-4 scale-90" data-sentry-element="svg" data-sentry-component="Mistral" data-sentry-source-file="mistral.tsx"><g clip-path="url(#_R_52aqansnmiv6hcpfivb_)" data-sentry-element="g" data-sentry-source-file="mistral.tsx"><path d="M54.3221 0H27.1531V27.0892H54.3221V0Z" fill="#FFD800" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M162.984 0H135.815V27.0892H162.984V0Z" fill="#FFD800" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M81.4823 27.0913H27.1531V54.1805H81.4823V27.0913Z" fill="#FFAF00" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M162.99 27.0913H108.661V54.1805H162.99V27.0913Z" fill="#FFAF00" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M162.972 54.168H27.1531V81.2572H162.972V54.168Z" fill="#FF8205" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M54.3221 81.2593H27.1531V108.349H54.3221V81.2593Z" fill="#FA500F" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M108.661 81.2593H81.4917V108.349H108.661V81.2593Z" fill="#FA500F" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M162.984 81.2593H135.815V108.349H162.984V81.2593Z" fill="#FA500F" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M81.4879 108.339H-0.00146484V135.429H81.4879V108.339Z" fill="#E10500" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path><path d="M190.159 108.339H108.661V135.429H190.159V108.339Z" fill="#E10500" data-sentry-element="path" data-sentry-source-file="mistral.tsx"></path></g><defs data-sentry-element="defs" data-sentry-source-file="mistral.tsx"><clipPath id="_R_52aqansnmiv6hcpfivb_" data-sentry-element="clipPath" data-sentry-source-file="mistral.tsx"><rect width="190.141" height="135" fill="white" data-sentry-element="rect" data-sentry-source-file="mistral.tsx"></rect></clipPath></defs></svg></div>
             <div><svg viewBox="0 0 24 24" width="24" height="24" xmlns="http://www.w3.org/2000/svg" className="pointer-events-none size-3.5 select-none sm:size-4 scale-125"><path d="M23.748 4.482c-.254-.124-.364.113-.512.234-.051.039-.094.09-.137.136-.372.397-.806.657-1.373.626-.829-.046-1.537.214-2.163.848-.133-.782-.575-1.248-1.247-1.548-.352-.156-.708-.311-.955-.65-.172-.241-.219-.51-.305-.774-.055-.16-.11-.323-.293-.35-.2-.031-.278.136-.356.276-.313.572-.434 1.202-.422 1.84.027 1.436.633 2.58 1.838 3.393.137.093.172.187.129.323-.082.28-.18.552-.266.833-.055.179-.137.217-.329.14a5.526 5.526 0 01-1.736-1.18c-.857-.828-1.631-1.742-2.597-2.458a11.365 11.365 0 00-.689-.471c-.985-.957.13-1.743.388-1.836.27-.098.093-.432-.779-.428-.872.004-1.67.295-2.687.684a3.055 3.055 0 01-.465.137 9.597 9.597 0 00-2.883-.102c-1.885.21-3.39 1.102-4.497 2.623C.082 8.606-.231 10.684.152 12.85c.403 2.284 1.569 4.175 3.36 5.653 1.858 1.533 3.997 2.284 6.438 2.14 1.482-.085 3.133-.284 4.994-1.86.47.234.962.327 1.78.397.63.059 1.236-.03 1.705-.128.735-.156.684-.837.419-.961-2.155-1.004-1.682-.595-2.113-.926 1.096-1.296 2.746-2.642 3.392-7.003.05-.347.007-.565 0-.845-.004-.17.035-.237.23-.256a4.173 4.173 0 001.545-.475c1.396-.763 1.96-2.015 2.093-3.517.02-.23-.004-.467-.247-.588zM11.581 18c-2.089-1.642-3.102-2.183-3.52-2.16-.392.024-.321.471-.235.763.09.288.207.486.371.739.114.167.192.416-.113.603-.673.416-1.842-.14-1.897-.167-1.361-.802-2.5-1.86-3.301-3.307-.774-1.393-1.224-2.887-1.298-4.482-.02-.386.093-.522.477-.592a4.696 4.696 0 011.529-.039c2.132.312 3.946 1.265 5.468 2.774.868.86 1.525 1.887 2.202 2.891.72 1.066 1.494 2.082 2.48 2.914.348.292.625.514.891.677-.802.09-2.14.11-3.054-.614zm1-6.44a.306.306 0 01.415-.287.302.302 0 01.2.288.306.306 0 01-.31.307.303.303 0 01-.304-.308zm3.11 1.596c-.2.081-.399.151-.59.16a1.245 1.245 0 01-.798-.254c-.274-.23-.47-.358-.552-.758a1.73 1.73 0 01.016-.588c.07-.327-.008-.537-.239-.727-.187-.156-.426-.199-.688-.199a.559.559 0 01-.254-.078c-.11-.054-.2-.19-.114-.358.028-.054.16-.186.192-.21.356-.202.767-.136 1.146.016.352.144.618.408 1.001.782.391.451.462.576.685.914.176.265.336.537.445.848.067.195-.019.354-.25.452z" fill="#4D6BFE"></path></svg></div>
             <div><svg height="1em" viewBox="0 0 24 24" width="1em" xmlns="http://www.w3.org/2000/svg" className="pointer-events-none size-3.5 select-none sm:size-4 scale-100"><title>Minimax</title><defs><linearGradient id="lobe-icons-minimax-fill" x1="0%" x2="100.182%" y1="50.057%" y2="50.057%"><stop offset="0%" stop-color="#E2167E"></stop><stop offset="100%" stop-color="#FE603C"></stop></linearGradient></defs><path d="M16.278 2c1.156 0 2.093.927 2.093 2.07v12.501a.74.74 0 00.744.709.74.74 0 00.743-.709V9.099a2.06 2.06 0 012.071-2.049A2.06 2.06 0 0124 9.1v6.561a.649.649 0 01-.652.645.649.649 0 01-.653-.645V9.1a.762.762 0 00-.766-.758.762.762 0 00-.766.758v7.472a2.037 2.037 0 01-2.048 2.026 2.037 2.037 0 01-2.048-2.026v-12.5a.785.785 0 00-.788-.753.785.785 0 00-.789.752l-.001 15.904A2.037 2.037 0 0113.441 22a2.037 2.037 0 01-2.048-2.026V18.04c0-.356.292-.645.652-.645.36 0 .652.289.652.645v1.934c0 .263.142.506.372.638.23.131.514.131.744 0a.734.734 0 00.372-.638V4.07c0-1.143.937-2.07 2.093-2.07zm-5.674 0c1.156 0 2.093.927 2.093 2.07v11.523a.648.648 0 01-.652.645.648.648 0 01-.652-.645V4.07a.785.785 0 00-.789-.78.785.785 0 00-.789.78v14.013a2.06 2.06 0 01-2.07 2.048 2.06 2.06 0 01-2.071-2.048V9.1a.762.762 0 00-.766-.758.762.762 0 00-.766.758v3.8a2.06 2.06 0 01-2.071 2.049A2.06 2.06 0 010 12.9v-1.378c0-.357.292-.646.652-.646.36 0 .653.29.653.646V12.9c0 .418.343.757.766.757s.766-.339.766-.757V9.099a2.06 2.06 0 012.07-2.048 2.06 2.06 0 012.071 2.048v8.984c0 .419.343.758.767.758.423 0 .766-.339.766-.758V4.07c0-1.143.937-2.07 2.093-2.07z" fill="url(#lobe-icons-minimax-fill)" fill-rule="nonzero"></path></svg></div>
