+++ src/components/LeaderboardActionsMenu.tsx
@@ -39,7 +39,7 @@ const LeaderboardActionsMenu: React.FC<LeaderboardActionsMenuProps> = ({
       {/* Hamburger Menu Button */}
       <button
         onClick={() => setIsOpen(!isOpen)}
-        className="p-1.5 text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 rounded-md transition-colors duration-200"
+        className="p-1.5 text-neutral-500 dark:text-neutral-400 hover:text-neutral-700 dark:hover:text-neutral-300 hover:bg-neutral-100 dark:hover:bg-neutral-700 rounded-md transition-colors duration-200"
         title="Actions menu"
       >
         <Menu className="w-4 h-4" />
@@ -47,7 +47,7 @@ const LeaderboardActionsMenu: React.FC<LeaderboardActionsMenuProps> = ({
 
       {/* Dropdown Menu */}
       {isOpen && (
-        <div className="absolute right-0 top-full mt-1 w-48 bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 rounded-lg shadow-lg z-50">
+        <div className="absolute right-0 top-full mt-1 w-48 bg-white dark:bg-neutral-800 border border-neutral-200 dark:border-neutral-700 rounded-lg shadow-lg z-50">
           <div className="py-1">
             {/* Share Button */}
             <button
@@ -55,7 +55,7 @@ const LeaderboardActionsMenu: React.FC<LeaderboardActionsMenuProps> = ({
                 onShare();
                 setIsOpen(false);
               }}
-              className="w-full px-4 py-2 text-left text-sm text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 flex items-center transition-colors duration-200"
+              className="w-full px-4 py-2 text-left text-sm text-neutral-700 dark:text-neutral-300 hover:bg-neutral-100 dark:hover:bg-neutral-700 flex items-center transition-colors duration-200"
             >
               <Share2 className="w-4 h-4 mr-3" />
               Share Leaderboard
