+++ src/components/SelectionReviewCard.tsx
@@ -41,64 +41,6 @@ export default function SelectionReviewCard({ challenge, template }: { challenge
       {/* Collapsible Content Section */}
       {isExpanded && (
         <div className="p-6 space-y-6 animate-in slide-in-from-top-2 duration-200">
-          {/* Main Selection Overview */}
-          <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
-            {/* Challenge Info */}
-            <div className="bg-white dark:bg-neutral-800 rounded-xl p-4 border border-blue-200 dark:border-blue-800 shadow-sm">
-              <div className="flex items-start gap-3">
-                <div className="w-10 h-10 bg-blue-100 dark:bg-blue-900/30 rounded-lg flex items-center justify-center flex-shrink-0">
-                  <Target className="w-5 h-5 text-blue-600 dark:text-blue-400" />
-                </div>
-                <div className="min-w-0 flex-1">
-                  <p className="text-xs font-semibold text-blue-600 dark:text-blue-400 uppercase tracking-wide mb-1">Challenge</p>
-                  <h3 className="text-base font-semibold text-neutral-800 dark:text-neutral-100 mb-1">
-                    {challenge.challenge_name}
-                  </h3>
-                </div>
-              </div>
-            </div>
-
-            {/* Template Info */}
-            <div className="bg-white dark:bg-neutral-800 rounded-xl p-4 border border-purple-200 dark:border-purple-800 shadow-sm">
-              <div className="flex items-start gap-3">
-                <div className="w-10 h-10 bg-purple-100 dark:bg-purple-900/30 rounded-lg flex items-center justify-center flex-shrink-0">
-                  <FileText className="w-5 h-5 text-purple-600 dark:text-purple-400" />
-                </div>
-                <div className="min-w-0 flex-1">
-                  <p className="text-xs font-semibold text-purple-600 dark:text-purple-400 uppercase tracking-wide mb-1">Template</p>
-                  <h3 className="text-base font-semibold text-neutral-800 dark:text-neutral-100 mb-1">
-                    {template.template_name}
-                  </h3>
-                </div>
-              </div>
-            </div>
-
-            {/* Language Pair */}
-            <div className="bg-white dark:bg-neutral-800 rounded-xl p-4 border border-emerald-200 dark:border-emerald-800 shadow-sm">
-              <div className="flex items-start gap-3">
-                <div className="w-10 h-10 bg-emerald-100 dark:bg-emerald-900/30 rounded-lg flex items-center justify-center flex-shrink-0">
-                  <Languages className="w-5 h-5 text-emerald-600 dark:text-emerald-400" />
-                </div>
-                <div className="min-w-0 flex-1">
-                  <p className="text-xs font-semibold text-emerald-600 dark:text-emerald-400 uppercase tracking-wide mb-1">Languages</p>
-                  <div className="flex items-center gap-2 mt-2">
-                    <div className="flex items-center gap-2 px-3 py-1.5 bg-emerald-50 dark:bg-emerald-900/20 rounded-lg border border-emerald-200 dark:border-emerald-800">
-                      <span className="text-sm font-medium text-emerald-700 dark:text-emerald-300">
-                        {challenge.from_language}
-                      </span>
-                    </div>
-                    <ArrowRight className="w-4 h-4 text-neutral-400 dark:text-neutral-500" />
-                    <div className="flex items-center gap-2 px-3 py-1.5 bg-purple-50 dark:bg-purple-900/20 rounded-lg border border-purple-200 dark:border-purple-800">
-                      <span className="text-sm font-medium text-purple-700 dark:text-purple-300">
-                        {challenge.to_language}
-                      </span>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-
           {/* Template Preview */}
           <div className="bg-white dark:bg-neutral-800 rounded-xl border border-neutral-200 dark:border-neutral-600 overflow-hidden">
             <div className="px-4 py-3 bg-neutral-50 dark:bg-neutral-700/50 border-b border-neutral-200 dark:border-neutral-600">
