+++ src/pages/Profile.tsx
@@ -7,10 +7,10 @@ const Profile = () => {
 
   if (isLoading) {
     return (
-      <div className="min-h-screen bg-gray-50 dark:bg-gray-700 flex items-center justify-center">
+      <div className="min-h-screen bg-neutral-50 dark:bg-neutral-900 flex items-center justify-center">
         <div className="text-center">
-          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-600 mx-auto"></div>
-          <p className="mt-4 text-gray-600 dark:text-gray-400">
+          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-primary-600 dark:border-primary-400 mx-auto"></div>
+          <p className="mt-4 text-neutral-600 dark:text-neutral-400">
             Loading profile...
           </p>
         </div>
@@ -22,27 +22,27 @@ const Profile = () => {
     <ProtectedRoute>
       <div className="py-0">
         <div className="max-w-4xl mx-auto px-4">
-          <div className="bg-white dark:bg-gray-800 rounded-lg shadow-md overflow-hidden">
+          <div className="bg-white dark:bg-neutral-800 rounded-lg shadow-md overflow-hidden">
             {/* Header */}
-            <div className="bg-gradient-to-r from-blue-600 to-blue-700 px-6 py-8">
+            <div className="bg-gradient-to-r from-primary-600 to-primary-700 dark:from-primary-500 dark:to-primary-600 px-6 py-8">
               <div className="flex items-center space-x-6">
                 {user?.picture ? (
                   <img
-                    className="w-20 h-20 rounded-full object-cover border-4 border-white shadow-lg"
+                    className="w-20 h-20 rounded-full object-cover border-4 border-white dark:border-neutral-700 shadow-lg"
                     src={user.picture}
                     alt={user.name || "User avatar"}
                   />
                 ) : (
-                  <div className="w-20 h-20 bg-white dark:bg-gray-800 rounded-full flex items-center justify-center border-4 border-white shadow-lg">
-                    <User className="w-10 h-10 text-gray-400 dark:text-gray-500" />
+                  <div className="w-20 h-20 bg-white dark:bg-neutral-800 rounded-full flex items-center justify-center border-4 border-white dark:border-neutral-700 shadow-lg">
+                    <User className="w-10 h-10 text-neutral-400 dark:text-neutral-500" />
                   </div>
                 )}
                 <div className="text-white">
                   <h1 className="text-3xl font-bold">{user?.name || "User"}</h1>
-                  <p className="text-blue-100 dark:text-blue-300 mt-1">
+                  <p className="text-primary-100 dark:text-primary-200 mt-1">
                     {user?.email}
                   </p>
-                  <div className="flex items-center mt-2 text-blue-100 dark:text-blue-300">
+                  <div className="flex items-center mt-2 text-primary-100 dark:text-primary-200">
                     <Calendar className="w-4 h-4 mr-2" />
                     <span className="text-sm">
                       Joined{" "}
@@ -66,10 +66,10 @@ const Profile = () => {
                       Basic Information
                     </h2>
                     <div className="space-y-3">
-                      <div className="flex items-center p-3 bg-gray-50 dark:bg-gray-700 rounded-lg">
-                        <Mail className="w-5 h-5 text-gray-400 dark:text-gray-500 mr-3" />
+                      <div className="flex items-center p-3 bg-neutral-50 dark:bg-neutral-700 rounded-lg">
+                        <Mail className="w-5 h-5 text-neutral-400 dark:text-neutral-500 mr-3" />
                         <div>
-                          <p className="text-sm text-gray-500 dark:text-gray-400">
+                          <p className="text-sm text-neutral-500 dark:text-neutral-400">
                             Email
                           </p>
                           <p className="font-medium text-neutral-600 dark:text-neutral-100">
@@ -78,10 +78,10 @@ const Profile = () => {
                         </div>
                       </div>
 
-                      <div className="flex items-center p-3 bg-gray-50 dark:bg-gray-700 rounded-lg">
-                        <User className="w-5 h-5 text-gray-400 dark:text-gray-500 mr-3" />
+                      <div className="flex items-center p-3 bg-neutral-50 dark:bg-neutral-700 rounded-lg">
+                        <User className="w-5 h-5 text-neutral-400 dark:text-neutral-500 mr-3" />
                         <div>
-                          <p className="text-sm text-gray-500 dark:text-gray-400">
+                          <p className="text-sm text-neutral-500 dark:text-neutral-400">
                             Full Name
                           </p>
                           <p className="font-medium text-neutral-600 dark:text-neutral-100">
@@ -90,10 +90,10 @@ const Profile = () => {
                         </div>
                       </div>
 
-                      <div className="flex items-center p-3 bg-gray-50 dark:bg-gray-700 rounded-lg">
-                        <Shield className="w-5 h-5 text-gray-400 dark:text-gray-500 mr-3" />
+                      <div className="flex items-center p-3 bg-neutral-50 dark:bg-neutral-700 rounded-lg">
+                        <Shield className="w-5 h-5 text-neutral-400 dark:text-neutral-500 mr-3" />
                         <div>
-                          <p className="text-sm text-gray-500 dark:text-gray-400">
+                          <p className="text-sm text-neutral-500 dark:text-neutral-400">
                             Account Status
                           </p>
                           <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 dark:bg-green-900/20 text-green-800 dark:text-green-400">
@@ -113,14 +113,14 @@ const Profile = () => {
                       Activity Summary
                     </h2>
                     <div className="space-y-3">
-                      <div className="flex items-center justify-between p-3 bg-blue-50 dark:bg-blue-900/20 rounded-lg">
+                      <div className="flex items-center justify-between p-3 bg-primary-50 dark:bg-primary-900/20 rounded-lg">
                         <div className="flex items-center">
-                          <FileText className="w-5 h-5 text-blue-600 dark:text-blue-400 mr-3" />
+                          <FileText className="w-5 h-5 text-primary-600 dark:text-primary-400 mr-3" />
                           <span className="font-medium text-neutral-600 dark:text-neutral-100">
                             Total Submissions
                           </span>
                         </div>
-                        <span className="text-2xl font-bold text-blue-600 dark:text-blue-400">
+                        <span className="text-2xl font-bold text-primary-600 dark:text-primary-400">
                           0
                         </span>
                       </div>
